requirements.txt
numpy
tifffile
treelib
pandas
requests
meshio
click
rich
tqdm>=4.46.1
bg-space
allensdk
sphinx
brainio>=0.0.16
vtkplotter
recommonmark
sphinx_rtd_theme
pydoc-markdown
black
pytest-cov
pytest
gitpython
coverage
pre-commit
PyMCubes